WIP (POOLER-70) Refactor out VSphere to separate class
WIP
